Update multiple tables of institution code
Create tables in your local database first. Open the sql file, search for "institution_code_1" and replace all with "institution_code_x", x being the id of the institution you want to create table for. 
f you want to populate the table, change the value of $inst in the php file and run the file. This should be replaced with automatic update in the future